refactor(hub): split residual hub.go into thematic files
hub.go now holds only the composition root and lifecycle: the Server
struct, NewServer (store/db -> services -> background loops wiring),
Run(ctx)/registerRoutes, and the serverAuth adapter for the httpserver
route table (281 LOC, target <500).

The remaining ~5.3k LOC moved verbatim (same package, no behavior
changes) into files by theme:

- auth.go: auth middlewares, token/tenant resolution, login/session
- webui.go: embedded web UI serving, branding/config endpoints
- claw_handlers.go: claw collection/detail HTTP handlers
- message_handlers.go: conversation message/activity handlers
- userws.go: user WebSocket conns, broadcast fan-out, JSON helpers
- provision.go: provider-agnostic provisioning + bootstrap status
- provision_daytona.go / provision_replicated.go / provision_exedev.go
  / provision_docker.go: per-provider provisioning and teardown
- background.go: provider status polling, status watchdog
- wake.go: wake messages and the initial-plan protocol
- clawenv.go: env/shell-script builders injected into sandboxes
- ssh.go: SSH exec/file-write helpers for remote VMs

All files are now under the 1,500 LOC phase-2 ceiling.
